refactor: 💡 moved pagination as component and other amends (#63)
* refactor: 💡 moved pagination as component and other amends

* chore: 🤖 target beta release
